#2773 - Refactoring request: remove warnings from build logs for linters * Refactor Example project and fix warning notifications * Set severity level for "no-explicit-any" to "error" * Remove unused operations from packages/ketcher-core/src/application/operations * Refactor Ketcher-core project and fix warning notifications whenever it's possible * Refactor Ketcher-standalone project and fix warning notifications whenever it's possible * Refactor Ketcher-polymer-editor-react project and fix warning notifications whenever it's possible
