+++ packages/ketcher-react/src/script/editor/tool/rotate-controller.ts
@@ -4,6 +4,7 @@ import Editor from '../Editor'
 import { getGroupIdsFromItemArrays } from './helper/getGroupIdsFromItems'
 import RotateTool from './rotate'
 import SelectTool from './select'
+import { getDifference, rotatePoint } from './rotate-controller.utils'
 
 type RaphaelElement = {
   [key: string]: any
@@ -797,36 +798,4 @@ class RotateController {
 }
 
 export default RotateController
-
-const rotatePoint = (centerPoint: Vec2, startPoint: Vec2, angle: number) => {
-  const oCenter = centerPoint
-  const oStart = startPoint
-
-  const centerStart = oStart.sub(oCenter)
-  const centerEnd = centerStart.rotate(angle)
-
-  const oEnd = oCenter.add(centerEnd)
-  return oEnd
-}
-
-export const getDifference = (
-  currentDegree: number,
-  structRotateDegree: number
-) => {
-  let abs = 0
-
-  // HACK: https://github.com/epam/ketcher/pull/2574#issuecomment-1539509046
-  if (structRotateDegree > 90) {
-    const positiveCurrentDegree =
-      currentDegree < 0 ? currentDegree + 360 : currentDegree
-    abs = Math.abs(positiveCurrentDegree - structRotateDegree)
-  } else if (structRotateDegree < -90) {
-    const negativeCurrentDegree =
-      currentDegree > 0 ? currentDegree - 360 : currentDegree
-    abs = Math.abs(negativeCurrentDegree - structRotateDegree)
-  } else {
-    abs = Math.abs(currentDegree - structRotateDegree)
-  }
-
-  return abs
-}
+export { getDifference }
